[MOV] mail, *: use PopoverView for Message/notificationIcon popover
