+++ addons/snailmail/static/src/components/snailmail_notification_popover_content_view/snailmail_notification_popover_content_view.js
@@ -1,0 +1,24 @@
+/** @odoo-module **/
+
+import { registerMessagingComponent } from '@mail/utils/messaging_component';
+
+const { Component } = owl;
+
+export class SnailmailNotificationPopoverContentView extends Component {
+
+    /**
+     * @returns {SnailmailNotificationPopoverContentView}
+     */
+    get snailmailNotificationPopoverContentView() {
+        return this.props.record;
+    }
+
+}
+
+Object.assign(SnailmailNotificationPopoverContentView, {
+    props: { record: Object },
+    template: 'snailmail.SnailmailNotificationPopoverContentView',
+});
+
+registerMessagingComponent(SnailmailNotificationPopoverContentView);
+
