+++ addons/snailmail/static/src/components/snailmail_notification_popover_content_view/snailmail_notification_popover_content_view.scss
