+++ addons/snailmail/static/src/components/snailmail_notification_popover_content_view/snailmail_notification_popover_content_view.xml
@@ -3,8 +3,8 @@
 
     <t t-name="snailmail.SnailmailNotificationPopoverContentView" owl="1">
         <div class="o_SnailmailNotificationPopoverContentView m-2" t-attf-class="{{ className }}" t-ref="root">
-            <i class="o_SnailmailNotificationPopoverContentView_icon" t-att-class="iconClass" role="img"/>
-            <span t-esc="iconTitle"/>
+            <i class="o_SnailmailNotificationPopoverContentView_icon" t-att-class="snailmailNotificationPopoverContentView.iconClass" role="img"/>
+            <span t-esc="snailmailNotificationPopoverContentView.iconTitle"/>
         </div>
     </t>
